sft run
