+++ realhf/impl/dataset/prompt_answer_dataset.py
@@ -37,6 +37,7 @@ class PromptAnswerDataset(torch.utils.data.Dataset):
         data = data_api.load_shuffle_split_dataset(util, dataset_path, dataset_builder)
 
         seqs = [x["prompt"] + x["answer"] + tokenizer.eos_token for x in data]
+        self.ids = [x["id"] for x in data]
         prompts = [x["prompt"] for x in data]
 
         self.tokens = tokenizer(
@@ -99,7 +100,7 @@ class PromptAnswerDataset(torch.utils.data.Dataset):
             keys=list(d.keys()),
             trailing_shapes={k: () for k in d.keys()},
             dtypes={k: d[k].dtype for k in d.keys()},
-            ids=[idx],
+            ids=[self.ids[idx]],
             seqlens={k: [seqlen] for k in d.keys()},
             data=d,
         )
