+++ realhf/impl/dataset/prompt_dataset.py
@@ -36,6 +36,7 @@ class PromptDataset(torch.utils.data.Dataset):
         data = data_api.load_shuffle_split_dataset(util, dataset_path, dataset_builder)
 
         prompts_str = [x["prompt"] for x in data]
+        self.ids = [x["id"] for x in data]
         util.tokenizer.padding_side = "left"
         prompt_encodings = util.tokenizer(
             prompts_str,
@@ -64,7 +65,7 @@ class PromptDataset(torch.utils.data.Dataset):
             keys=["packed_prompts"],
             trailing_shapes=dict(packed_prompts=()),
             dtypes=dict(packed_prompts=torch.long),
-            ids=[idx],
+            ids=[self.ids[idx]],
             seqlens=dict(packed_prompts=[torch.tensor([self.prompt_lengths[idx]])]),
             data=dict(packed_prompts=torch.tensor(self.prompts[idx], dtype=torch.long)),
         )
